requirements.txt
# ==============================================================================
# AltsReimaginedAPI - Python Dependencies
# ==============================================================================
# Install with: pip install -r requirements.txt
# ==============================================================================

# ------------------------------------------------------------------------------
# Web Framework
# ------------------------------------------------------------------------------
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
PyJWT>=2.8.0
python-multipart>=0.0.6

# ------------------------------------------------------------------------------
# Database & ORM
# ------------------------------------------------------------------------------
sqlalchemy>=2.0.0
alembic>=1.12.0

# PostgreSQL Driver
psycopg2-binary>=2.9.0

# SQL Server Driver (uncomment if using SQL Server)
pyodbc>=5.0.0

# ------------------------------------------------------------------------------
# Configuration & Environment
# ------------------------------------------------------------------------------
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ------------------------------------------------------------------------------
# Logging & Monitoring (Optional)
# ------------------------------------------------------------------------------
# Azure Application Insights (uncomment if using App Insights logging)
opencensus-ext-azure>=1.1.0
# applicationinsights>=0.11.0

# ------------------------------------------------------------------------------
# Development Dependencies
# ------------------------------------------------------------------------------
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# httpx>=0.24.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
